requirements.txt
alpaca-trade-api==0.42
astroid==2.2.5
asyncio-nats-client==0.10.0
awscli==1.18.19
botocore==1.15.19
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Deprecated==1.2.7
docutils==0.15.2
Flask==1.1.1
Flask-Debug==0.4.3
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.5
joblib==0.14.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mysql-connector-python==8.0.18
mysqlclient==1.4.2.post1
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
pandas-datareader==0.8.1
protobuf==3.11.2
pyasn1==0.4.8
pylint==2.3.1
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
schedule==0.6.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.12.0
textblob==0.15.3
tweepy==3.8.0
typed-ast==1.4.0
urllib3==1.24.3
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
websocket-client==0.56.0
websockets==8.1
Werkzeug==0.15.5
wrapt==1.11.2
